Arena Generation from Dropdown menu.
You can now change the arena type from the drop down menu in the main
arena page. As of now it regenerates the bot population when it
re-draws it (I’m working on how to pass the current bots at the
moment). I added the case for random arenas which can be 3 to 20 sides.
Some minor tweaks in arena.js were also done.