Update Commit
Lost the select and edit changes in previous commits, this is a working commit of a branch that has the select
and edit changes so we can get them back in master.

Other changes may have been lost, but at this point they will need to be inserted back in manually.